+++ bmp_manipulaitor.cbp
@@ -11,6 +11,7 @@
 				<Option object_output="obj/Debug/" />
 				<Option type="1" />
 				<Option compiler="gcc" />
+				<Option parameters="--tonalidad-azul=70 resources/imagen.bmp --escala-de-grises resources/imagen.bmp" />
 				<Compiler>
 					<Add option="-g" />
 				</Compiler>
@@ -35,6 +36,8 @@
 			<Option compilerVar="CC" />
 		</Unit>
 		<Unit filename="funciones_grupo.h" />
+		<Unit filename="funciones_quiroga.cpp" />
+		<Unit filename="funciones_quiroga.h" />
 		<Unit filename="funciones_zapata.c">
 			<Option compilerVar="CC" />
 		</Unit>
