+++ funciones_quiroga.h
@@ -1,0 +1,9 @@
+#ifndef FUNCIONES_QUIROGA_H_INCLUDED
+#define FUNCIONES_QUIROGA_H_INCLUDED
+
+#include "funciones_grupo.h"
+
+
+void concatenarImagenVertical (PixelRGB *imagenIzquierda, const BmpHeader *imgIzquierdaHeader,PixelRGB *imagenDerecha, const BmpHeader *imgDerechaHeader);
+
+#endif // FUNCIONES_QUIROGA_H_INCLUDED
