+++ funciones_zapata.h
@@ -10,6 +10,5 @@ void convertirEscalaDeGrises(PixelRGB *pixels, const BmpHeader *imgHeader);
 void cambioTonalidad(PixelRGB *imagenOriginal, const BmpHeader *imgHeader, float rojo, float verde, float azul);
 void testHeader(BmpHeader* h);
 void parse_argv(int argc, char* argv[], OpcionesImagen *opciones);
-void ejecutarOperaciones(OpcionesImagen *opciones);
 
 #endif // FUNCIONES_ZAPATA_H
